+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Adding the `phonetics/onca` namespace.
 * Adding the `structures/bk-tree` namespace.
 * Adding the `structures/symspell` namespace.
+* Adding the `structures/vp-tree` namespace.
 * Adding the `sampler` options to `clustering/k-means`.
 * Adding the `stats/descriptive#.quantile` function.
 * Adding the `stats/descriptive#.median` function.